+++ changelog.md
@@ -1,4 +1,5 @@
-# Cosmic Core 0.6.7
-Engine.
-
+# Cosmic Core 0.7.4
 # DO NOT MANUALLY UPDATE COSMIC CORE FOR FRONTIERS **I WILL NOT WORK**
+
+    Added Linked Terminal
+    Added Drone Station, Drones, and Drone Maint Hatch (In Testing)